+++ tests/testthat/test_smoke.R
@@ -52,6 +52,7 @@ test_that("file generation works", {
   expect_true(file.exists(file.path(output_dir, 'Charts.docx')))
   expect_true(file.exists(file.path(output_dir, 'nearest_neighbors.csv')))
   expect_true(file.exists(file.path(output_dir, 'count_within.csv')))
+  expect_true(file.exists(file.path(output_dir, 'session_info.txt')))
 
   # Check numbers against known good